+++ libs/tailwind/src/schematics/nx-setup/nx-setup.ts
@@ -6,17 +6,9 @@ import {
 } from '@angular-devkit/schematics';
 import {
   addDepsToPackageJson,
-  getProjectGraphFromHost,
   getWorkspace,
-  InsertChange,
-  ProjectGraph,
-  ProjectGraphNode,
-  projectRootDir,
-  ProjectType,
   updateWorkspace,
 } from '@nrwl/workspace';
-import { appsDir, libsDir } from '@nrwl/workspace/src/utils/ast-utils';
-import { DEPENDENCIES } from '../../constants';
 import {
   addConfigFiles,
   getLatestNodeVersion,
@@ -24,16 +16,9 @@ import {
   updateProjectRootStyles,
   updateWorkspaceTargets,
 } from '../../utils';
+import { normalizeOptionsNx } from '../../utils/normalize-options-nx';
+import { updateIndexHtml } from '../../utils/update-index-html';
 import type { TailwindSchematicsOptions } from '../schema';
-
-interface NormalizedTailwindSchematicsOptions
-  extends TailwindSchematicsOptions {
-  projectName: string;
-  projectRoot: string;
-  projectDirectory: string;
-  appsDir?: string;
-  libsDir?: string;
-}
 
 export default function (options: TailwindSchematicsOptions): Rule {
   return (tree, context) => {
@@ -44,26 +29,31 @@ export default function (options: TailwindSchematicsOptions): Rule {
       return;
     }
 
-    const { enableTailwindInComponentsStyles, projectName, appsDir, libsDir } = normalizeOptions(
-      options,
-      tree,
-      context
-    );
+    const normalizedOptions = normalizeOptionsNx(options, tree, context);
 
     return chain([
-      addDependenciesToPackageJson(),
-      addConfigFiles(enableTailwindInComponentsStyles, appsDir, libsDir),
-      updateWorkspaceTargets(projectName, updateWorkspace),
-      updateProjectRootStyles(projectName, getWorkspace, updateWorkspace),
+      addDependenciesToPackageJson(normalizedOptions.dependencies),
+      addConfigFiles(normalizedOptions),
+      updateWorkspaceTargets(normalizedOptions.projectName, updateWorkspace),
+      updateProjectRootStyles(
+        normalizedOptions.projectName,
+        getWorkspace,
+        updateWorkspace
+      ),
+      updateIndexHtml(
+        normalizedOptions.projectName,
+        normalizedOptions.darkMode,
+        updateWorkspace
+      ),
     ])(tree, context);
   };
 }
 
-function addDependenciesToPackageJson(): Rule {
+function addDependenciesToPackageJson(dependencies: string[]): Rule {
   return async (tree: Tree, ctx: SchematicContext) => {
     const devDeps = (
       await Promise.all(
-        [...DEPENDENCIES].map((dep) =>
+        dependencies.map((dep) =>
           getLatestNodeVersion(dep).then(({ name, version }) => {
             ctx.logger.info(`✅️ Added ${name}@${version}`);
             return { name, version };
@@ -77,48 +67,4 @@ function addDependenciesToPackageJson(): Rule {
 
     return addDepsToPackageJson({}, devDeps)(tree, ctx) as Rule;
   };
-}
-
-function normalizeOptions(
-  options: TailwindSchematicsOptions,
-  tree: Tree,
-  context: SchematicContext
-): NormalizedTailwindSchematicsOptions {
-  const project = getDefaultProjectFromGraph(
-    getProjectGraphFromHost(tree),
-    options.project
-  );
-
-  if (project == null) {
-    const msg = `Cannot find any Angular project in the current workspace.`;
-    context.logger.fatal(msg);
-    throw new Error(msg);
-  }
-
-  return {
-    ...options,
-    project: project.name,
-    projectName: project.name,
-    projectDirectory: project.data.root
-      .split(projectRootDir(ProjectType.Application) + '/')
-      .pop(),
-    projectRoot: project.data.root,
-    appsDir: appsDir(tree),
-    libsDir: libsDir(tree),
-  };
-}
-
-function getDefaultProjectFromGraph(
-  graph: ProjectGraph,
-  projectName?: string
-): ProjectGraphNode {
-  if (projectName) return graph.nodes[projectName];
-  return Object.values(graph.nodes).find(
-    (node) =>
-      node.type === 'app' &&
-      node.data.projectType === ProjectType.Application &&
-      Object.values(node.data.architect).some((target) =>
-        target.builder.includes('angular')
-      )
-  );
 }